Gallery Plugin
This plugin is called by adding the markup $! gallery !$ to a markdown page
or post file.  Images will be found in the gallery/NAME directory, where
NAME is the basename of the page or post.  All files (jpg) in that directory
will be processed, thumbnails generated, and HTML pages with prev, next
and up links generated.  All thumbnails are then displayed at the place
where the gallery plugin was called.